Post merge fixes
Source: https://github.com/msupply-foundation/open-msupply/commit/9904c58b8078a07f05c486102ddeb2aa150b28f3

+++ server/service/src/invoice_line/stock_out_line/insert/mod.rs
@@ -21,7 +21,7 @@ pub struct InsertStockOutLine {
     pub note: Option<String>,
     pub location_id: Option<String>,
     pub batch: Option<String>,
-    pub pack_size: Option<i32>,
+    pub pack_size: Option<f64>,
     pub expiry_date: Option<NaiveDate>,
     pub cost_price_per_pack: Option<f64>,
     pub sell_price_per_pack: Option<f64>,
